backend/requirements.txt
aiomysql==0.0.20
aiosqlite==0.15.0
asynctest==0.13.0
atomicwrites==1.4.0
attrs==20.2.0
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.1.1
ecdsa==0.14.1
fastapi==0.61.1
gunicorn==20.0.4
h11==0.9.0
httpcore==0.11.0
httpx==0.15.4
idna==2.10
iniconfig==1.0.1
iso8601==0.1.13
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.5.0
packaging==20.4
passlib==1.7.2
pluggy==0.13.1
prompt-toolkit==3.0.7
py==1.9.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.6.1
PyJWT==1.7.1
PyMySQL==0.9.2
pyparsing==2.4.7
PyPika==0.39.1
pytest==6.0.2
python-dotenv==0.14.0
python-jose==3.2.0
python-multipart==0.0.5
python-rapidjson==0.9.1
requests==2.24.0
rfc3986==1.4.0
rsa==4.6
six==1.15.0
sniffio==1.1.0
starlette==0.13.6
toml==0.10.1
tortoise-orm==0.16.15
typing-extensions==3.7.4.3
urllib3==1.25.10
uvicorn==0.11.8
wcwidth==0.2.5
websockets==8.1
wincertstore==0.2
XlsxWriter==1.3.6